rename files and use our prettier config
